requirement.txt
adafruit-io==2.7.0
keras==2.11.0
opencv-python==4.7.0.68
numpy==1.24.2
pyserial==3.5
Pillow==9.4.0